arch-vega: SDWA for floating point dests
There is an issue using the current template for SDWA as it uses bitwise
operation to select sub-dword values. Therefore we add a special
template function for floating point SDWA destinations. It seems the
only possible values for SDWA dest sel are dword, so we only allow for
that. This means there are no unused bits, so that field can be ignored.
The only possibility is the clamp bit, which is implemented.